# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Admin Platform API
  version: "1.0.0"
  description: |
    Backend API for the Admin Platform.
    All module endpoints require **Bearer JWT** authentication and pass through a
    four-layer guard chain: `authenticateJwt -> requireRole -> requireFeatureEnabled -> requirePermission`.

    **Error envelope** (all non-2xx responses):
    ```json
    { "code": "<ErrorCode>", "message": "<human string>" }
    ```

    **Pagination envelope** (list endpoints):
    ```json
    { "items": [...], "page": 1, "limit": 50, "total": 123 }
    ```
  contact:
    name: Admin Platform
servers:
  - url: http://localhost:7002
    description: Local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Two-tier JWT.  Tokens are signed with role-specific secrets
        (`JWT_SECRET_SUPER_ADMIN` / `JWT_SECRET_ADMIN`).
        Payload: `{ sub, email, role }`.

  # ─── Reusable schemas ───────────────────────────────────────────────
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }

    # ── Auth ──
...
    Ticket:
      type: object
      properties:
        _id: { type: string }
        ticketNumber: { type: string }
        subject: { type: string }
        description: { type: string }
        requesterName: { type: string }
        requesterEmail: { type: string }
        channel: { type: string, enum: [email, chat, phone, web] }
        priority: { type: string, enum: [low, medium, high, urgent] }
        status: { type: string, enum: [open, in_progress, pending_customer, resolved, closed] }
        tags: { type: array, items: { type: string } }
        assignedToUserId: { type: string }
        firstResponseAt: { type: string, format: date-time }
        resolvedAt: { type: string, format: date-time }
        closedAt: { type: string, format: date-time }
        comments:
          type: array
          items:
            type: object
            properties:
              message: { type: string }
              authorUserId: { type: string }
              isInternal: { type: boolean }
              createdAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }

    # ── File Manager ──
...
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: MongoDB ObjectId

  # ─── Reusable responses ───────────────────────────────────────────
...
paths:
...
  /api/v1/support-tickets/tickets:
    get:
      tags: [Support Tickets]
      summary: List tickets (paginated)
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginationMeta"
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/Ticket" } }
    post:
      tags: [Support Tickets]
      summary: Create ticket
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
  /api/v1/support-tickets/tickets/{id}:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    patch:
      tags: [Support Tickets]
      summary: Update ticket (not closed)
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
    delete:
      tags: [Support Tickets]
      summary: Delete ticket (closed only)
      responses:
        "204": { description: Deleted }
  /api/v1/support-tickets/tickets/{id}/transition:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    post:
      tags: [Support Tickets]
      summary: Transition ticket status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to]
              properties:
                to: { type: string, enum: [open, in_progress, pending_customer, resolved, closed] }
                note: { type: string, maxLength: 1000 }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
  /api/v1/support-tickets/tickets/{id}/comments:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    post:
      tags: [Support Tickets]
      summary: Add comment (internal requires super_admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string, minLength: 1, maxLength: 4000 }
                isInternal: { type: boolean, default: false }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
...